Set up initial load for website
Fixed a bug where Index.html did not publish properly to the web server. Edited Index.html so that it automatically redirects to Login.aspx (the original publishing issue occurred when I tried to set Login.aspx as the startup page, so this seemed like the best workaround). If the page does not redirect after a period of time, text will appear with a link that the user can follow.