Added new pages, updated GUI, changed file type
Added the CurrentCommittees, LatestElectionResults, and UserLanding pages. Added some GUI elements to those pages. Fixed a bug in EditUser that was caused by the wrong SQL Data Source being selected. Updated web.config to allow custom error messages to be displayed on the client-side. Changed AdminLanding from an html file to an aspx file since it will need to access a session variable in the future.