Login Functionality
The website compares the input username and password against the database to see if the user has an account. If they do, it checks if they are an admin, user, or something else, and the website will redirect them to the correct webpage for that account type. There is also a link on the login page for easy access to the other pages.